Split All Commands Over Multiple Files
